pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "pyventory"
version = "3.3.2"
description = "Ansible Inventory implementation that uses Python syntax"
authors = ["Serge Matveenko <lig@countzero.co>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/lig/pyventory"
repository = "https://github.com/lig/pyventory.git"
documentation = "https://readthedocs.org/projects/pyventory/"
keywords = ["devops", "ansible", "inventory", "terraform", "vars"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: System :: Systems Administration",
]

[tool.poetry.dependencies]
python = "^3.10"
attrs = "^22.1.0"
ordered-set = "^4.0.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
pytest-black = "^0.3.11"
pytest-flake8 = "^1.1.1"
pytest-isort = "^3.0.0"
pytest-mypy = "^0.10.0"
rstcheck = "^6.1.0"
Sphinx = "^5.2.3"

[tool.black]
skip-string-normalization = true

[tool.isort]
profile = "black"
lines_after_imports = 2
